Use heredocs in filter tests to avoid confusing escaping
The configuration snippets testing named capture replacement were
confusing because the double-backslashes are an artefact of ruby string
escaping.

This commit uses heredocs on those snippets to avoid having to use the
extra backslash to escape the string. This means that these snippets can
now be used as-is in documentation or fluentd configuration.